Fix ellipsis parsing, by adding whitepace match before ...
This took several debug sessions to find. Ellipsis worked at the
start of a line, and also when using regex directly, but not in
an example like
    1 ...
    + 2

It turns out it was not a problem with lexing presedence, nor with
the regex or escaping. Instead the problem is the space in front of
it, which is not part of the regex. Whitespace is only stripped in a
later lexer group, so needs to be done explicitly here.